refactor: move utils to separate folder
